chore(#42): modificando estrutura do backend
