+++ src/main/java/fatec/porygon/dto/LogDto.java
