+++ src/main/java/fatec/porygon/entity/Log.java
