+++ src/main/java/fatec/porygon/enums/StatusArea.java
