+++ src/main/java/fatec/porygon/service/LogService.java
